+++ src/groups_users_lib/managers/getters/user_getter.py
@@ -1,7 +1,7 @@
 """Get users from the shell."""
 from typing import AsyncIterable
 
-from shell_executor_lib import CommandManager
+from shell_executor_lib import CommandManager, CommandError
 from groups_users_lib import UserNotExistError
 
 
@@ -12,7 +12,7 @@ class UserGetter:
         """Initialize the UserGetter.
 
         Args:
-            command_manager: To make  commands in the shell.
+            command_manager: To make commands in the shell.
         """
         self._command_manager = command_manager
 
@@ -41,14 +41,20 @@ class UserGetter:
 
         Raises:
             UserExistError: If the user not exist.
-            CommandError: If the exit code is not 0.
+            CommandError: If the error is not excepted.
         """
-        data_user: list[str] = await self._command_manager.execute_command(
-            "/bin/cat /etc/passwd | /bin/awk -F : '{print \\$3,\\$1,\\$7,\\$6,\\$4}'" + f"| grep {user_name}", False)
+        try:
+            data_user: list[str] = await self._command_manager.execute_command(
+                "/bin/cat /etc/passwd | /bin/awk -F : '{print \\$3,\\$1,\\$7,\\$6,\\$4}'" + f"| grep {user_name}",
+                False)
 
-        if len(data_user) > 0:
-            data: list[str] = data_user[0].split(" ")
-            if len(data) > 2:
-                return int(data[0]), data[1], data[2], data[3], int(data[4])
+            if len(data_user) > 0:
+                data: list[str] = data_user[0].split(" ")
+                if len(data) > 2:
+                    return int(data[0]), data[1], data[2], data[3], int(data[4])
+
+        except CommandError as error:
+            if error.status_code != 1:
+                raise error
 
         raise UserNotExistError(user_name)