+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [0.1.2] - 2024-03-22
 ### Fixed
 - @javierfh03 - Fix incorrect exception when find nonexistent user.
+- @javierfh03 - Fix incorrect exception when find nonexistent group.
 ### Changed
 - @javierfh03 - Change group internal structure.